Refactored package name Fixed some schema issues with multivalued attributes
