+++ src/main/assembly/connector.xml
@@ -18,7 +18,6 @@
           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
           xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.0
                               http://maven.apache.org/xsd/assembly-1.1.0.xsd">
-
     <id>connector</id>
 
     <formats>
