+++ src/main/java/com/evolveum/polygon/notes/DominoAttribute.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.evolveum.openicf.lotus;
+package com.evolveum.polygon.notes;
 
 import org.identityconnectors.framework.common.objects.AttributeInfo;
