+++ src/main/resources/com/evolveum/polygon/notes/Message.properties
